ios: show server selector on new-session page, hide find-in-page, fix dismiss shadow flicker
## Summary

- Surface the iOS native server selector on the new-session landing
  screen, not just inside an active conversation. Extracted the
  visibility hook from `ChatPage` into a shared
  `useNativeServerSwitcher` module (avoids a circular import, since
  `ChatPage` already imports `NewChatLandingScreen`) and wired it into
  `NewChatLandingScreen` against the landing surface element.
- Removed the "Find in Page" item from the iOS `ServerSwitcher` menu and
  dropped the now-unused `WebViewModel.showFind()`.
- Fixed a jarring UX glitch where the selector pill lost its drop shadow
  for a beat after the menu was dismissed. The chrome
  (material/border/shadow) was inside the `Menu`'s `label:` closure, so
  UIKit's menu-presentation snapshot dropped the shadow layer during the
  open/dismiss morph. Moved that chrome onto the Menu's persistent host
  view so it survives the snapshot.

## Type of change

- [x] Bug fix
- [x] Feature
- [ ] Refactor / chore
- [ ] Docs
- [ ] Test / CI
- [ ] Breaking change

## Test coverage

- [ ] Unit tests added / updated
- [ ] Integration tests added / updated
- [ ] E2E tests added / updated
- [x] Manual verification completed
- [ ] Existing tests cover this change
- [ ] Not applicable

## Coverage rationale

Web side verified with `npm run type-check` (clean) and the existing
suites `npx vitest run src/lib/nativeBridge.test.ts` (23 passed) plus
`src/shell/NewChatDialog.test.tsx` and `NewChatDialog.flow.test.tsx`
(132 passed, 1 skipped). iOS changes verified by a full simulator build
(`xcodebuild ... build` -> BUILD SUCCEEDED); the shadow-flicker fix is a
visual/timing behavior not expressible as an automated test.
